Propogate error back to middleware when serving files
This PR bubbles the error back to the middleware that is supposed to serve the files. For example, serve-files middleware expects an optional `err` attribute and logs the error. If an asset is being requested to be served and is not present, the error will never be thrown. User will only know if opened the network tab in a developer options.

Testing:
1. Tested ti with a dummy app that requested for a non existent asset and got the error stacktrace thrown in the serve logs.

PS: This repo needs some serious love. Will send another PR to beatify the code.