Change: Errors page to use CDNs
I thought we were going to be able to inline everything but that's not
really feasible given the different languages we're going to want to
support.

However CDNs are also not acceptable since it requires an internet
connection.

For now we'll do this to ease the burden of development so that we don't
have to scroll through a bunch of code that we don't care about.