Add original Tomster to the Dang! error page for encouragement
For those of us who end up on the Ember build error page, a Tomster
to offer encouragement for the hamster wheel we're in at the moment.
I mentioned this to @wifelette in the Ember Slack, and though she
thought a few different Tomsters would work, I thought this one was
the kindest option.